Extract ModelStub class from AddCommandTest into its own file in testutil
